Improve tests and avoid divide-by-zero in func
In the numpy code returned by lambdify, you can evalute a divide-by-zero
in the Piecewise functions returned by limit().  This adds tests that
demonstrate this, and avoids it with an epsilon=1e-10 to prevent exact
division by zero.

In future proper avoided evaluation of the branches not taken would
avoid the need for this.

Also check for correct usage of model.eval() - it expects a symbol like
X[3] not an arbitrary expression.